.claude/skills/sales-deck/requirements.txt
python-pptx>=0.6.23
Pillow>=10.0
lxml>=5.0
